Problem: https://leetcode.com/problems/minimum-window-substring/
Find the minimum subarray string in a string which contains all the characters as in the second input.
O(n) solution by using the count of the number of characters already found against the number of characters yet to find.
In the worst case we travel the entire array twice.