@Problem: https://leetcode.com/problems/restore-ip-addresses/
Given a string "25525511135" . Find all the valid IP addresses for the string.
Solution works by taking breaking the string into 4 octets starting with 1 character each. If the size of the 4 octets equals the given input, then proceed to check if each of the octet is <=255. If yes, then add to the result.