@Random: https://leetcode.com/problems/linked-list-random-node/
Generate a Random node by traversing the linked list only once.
Based on the idea of Knuth Shuffle. Generate a random number between 0 and i-1 where i is the index of the node currently visiting. If the generated number equals i replaced the result node value with current node value. Every number is generated with even probability of 1/n.