@Problem: https://leetcode.com/problems/sort-characters-by-frequency/
Given a string, sort it in decreasing order of frequency of characters. Use a vector of strings. Group the characters as you traverse and store in the vector. The vector is of length 128. And then sort the vector according to the size and pop the strings which have size 0. Then concatenate all the strings in vector.